refactor(backend): move signup/login functions out of test files
